+++ src/Api/IriToIdentifierConverterInterface.php
@@ -14,20 +14,18 @@ declare(strict_types=1);
 namespace ApiPlatform\Core\Api;
 
 use ApiPlatform\Core\Exception\InvalidArgumentException;
-use ApiPlatform\Core\Exception\ItemNotFoundException;
-use ApiPlatform\Core\Exception\RuntimeException;
 
 /**
- * Converts a plain identifier to an IRI
+ * Converts a plain identifier to an IRI.
  *
  * @author Hamza Amrouche <hamza.simperfit@gmail.com>
  */
-interface IriPlainIdentifierAwareConverterInterface extends IriConverterInterface
+interface IriToIdentifierConverterInterface
 {
     /**
      * Gets the IRI associated with the given plain identifier.
      *
-     * @param string|int|array $id
+     * @param array $id
      * @param string           $resourceClass
      * @param int              $referenceType
      *
@@ -35,5 +33,5 @@ interface IriPlainIdentifierAwareConverterInterface extends IriConverterInterfac
      *
      * @return string
      */
-    public function getIriFromPlainIdentifier($id, string $resourceClass, int $referenceType = UrlGeneratorInterface::ABS_PATH): string;
+    public function getIriFromPlainIdentifier(array $id, string $resourceClass, int $referenceType = UrlGeneratorInterface::ABS_PATH): string;
 }