+++ tests/Fixtures/TestBundle/Entity/Dummy.php
@@ -171,6 +171,8 @@ class Dummy
     public function setId($id)
     {
         $this->id = $id;
+
+        return $this;
     }
 
     public function setName($name)